+++ MercutioEngine/meNavSystem.cpp
@@ -18,7 +18,7 @@ glm::vec3 NavSystem::GetNode(glm::vec3 pos)
 		if (forward) { navNode++; }
 		else { navNode--; }
 		if (navNode == navPath.size()) { navNode = 0; }
-		if (navNode == -1) { navNode = navPath.size() - 1; }
+		if (navNode == -1) { navNode = (int)navPath.size() - 1; }
 	}
 	return navPath[navNode];
 }
